requirements.txt
djangorestframework>=3.14,<4.0
celery>=5.2,<6.0
redis>=4.0,<6.0
python-dotenv>=1.0,<2.0
requests>=2.0,<3.0 
pytz
psycopg2-binary
flower